### 2023-04-06 #5345 @editorialbot
[REVIEW]: FrESCO: Framework for Exploring Scalable Computational Oncology
**Submitting author:** @aspannaus (<a href="http://orcid.org/0000-0002-2521-3657">Adam Spannaus</a>)
**Repository:** https://github.com/DOE-NCI-MOSSAIC/FrESCO
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v0.2.4
