### 2023-05-10 #5345 @SimonBiggs
Hi @aspannaus,

I can confirm that cloning is now working. But I cannot raise an issue there either:

![image](https://github.com/openjournals/joss-reviews/assets/6559099/5363e88d-707f-4127-97cf-73e62d00541c)
