+++ abstention/abstention.py
@@ -1,0 +1,490 @@
+"""
+    Module implementing deep-abstaining classifier.
+
+"""
+
+#    pylint: disable=C0121
+
+import copy
+
+import torch
+import torch.nn.functional as F
+
+import numpy as np
+# import numpy.ma as ma
+
+from sklearn.metrics import accuracy_score
+
+from torch import nn
+
+
+class DacLoss(nn.Module):
+    """
+        Custom loss class for the DAC.
+
+        Args:
+            loss_fin: torch.nn.loss fun, either Cross Entropy or
+                Binary Cross Entropy (for multilabel)
+            multilabel: bool, multilabel problem?
+            alphas: torch tensor of alphas for the DAC
+    """
+    def __init__(self, loss_fun, multilabel, alphas):
+        super().__init__()
+        self.loss_fun = loss_fun
+        self.multilabel = multilabel
+        self.alphas = alphas
+
+    def forward(self,
+                y_pred: torch.tensor,
+                y_true: torch.tensor,
+                idx: int,
+                ntask_abs_prob: int = 1) -> torch.tensor:
+        """Compute DAC loss.
+
+        Params:
+            y_pred - torch.tensor: logits from model.forward()
+            y_true - torch.tensor: integer values of ground truth
+            idx: int index associated with a task
+            ntask_abs_prob: probability of ntask abstaining on the document
+
+        Pre-condition: loss function defined in class constructor
+
+        """
+
+        eps = 1e-4
+
+        # updated loss
+        if self.multilabel:
+            log_probs = self._log_sigmoid(y_pred)
+        else:  # need the log_softmax to account of correct operations for the multilabel case
+            log_probs = F.log_softmax(y_pred, dim=1)
+
+        p_k = log_probs[:, :-1]
+        p_x = torch.exp(log_probs[:, -1])
+        om_px = 1.0 - p_x + eps
+
+        if self.multilabel:
+            tmp_loss = torch.mean(self.loss_fun(p_k, y_true), -1) + torch.log(om_px)
+        else:
+            tmp_loss = self.loss_fun(p_k, y_true) + torch.log(om_px)
+
+        loss = (om_px + 1 - ntask_abs_prob) * tmp_loss - self.alphas[idx] * torch.log(om_px)
+
+        return torch.mean(loss)
+
+    @staticmethod
+    def _log_sigmoid(x):
+        """Compute numerically stable log sigmoid function."""
+        x_max = torch.amax(x, keep_dim=True)
+        if not x_max.isfinite():
+            x_max = 0.0
+        tmp = x - x_max
+        exp_tmp = torch.exp(tmp)
+        s = torch.sum(exp_tmp, keepdim=True)
+        out = torch.log(s)
+        return tmp - out
+
+
+class AbstainingClassifier():
+    """Class for deep abastaining classifier.
+
+        Params:
+            id2label - dict : dict mapping int values to label values for each task
+            kw_args - dict: dictionary with necessary keywords for abstention
+            device - str: 'cuda' or 'cpu', set from model_suite.py
+    """
+    def __init__(self, kw_args: dict, device: torch.device, class_weights=None, clc: bool = False):
+
+        self.tasks = kw_args['data_kwargs']['tasks']
+        self.ntask = kw_args['abstain_kwargs']['ntask_flag']
+
+        if self.ntask:
+            self.ntask_tasks = kw_args['abstain_kwargs']['ntask_tasks']
+            self.ntask_alpha = kw_args['abstain_kwargs']['ntask_alpha']
+            self.ntask_max_abs = kw_args['abstain_kwargs']['ntask_max_abs']
[…]
