+++ README.md
@@ -1,0 +1,124 @@
+# FrESCO: Framework for Exploring Scalable Computational Oncology
+
+## Motivation
+The National Cancer Institute (NCI) monitors population level cancer trends as part of its Surveillance, Epidemiology, and End Results (SEER) program. This program consists of state or regional level cancer registries which collect, analyze, and annotate cancer pathology reports. From these annotated pathology reports, each individual registry aggregates cancer phenotype information and summary statistics about cancer prevalence to facilitate population level monitoring of cancer incidence. Extracting cancer phenotype from these reports is a labor intensive task, requiring specialized knowledge about the reports and cancer. Automating this information extraction process from cancer pathology reports has the potential to  improve not only the quality of the data by extracting information in a consistent manner across registries, but to improve the quality of patient outcomes by reducing the time to assimilate new data and enabling time-sensitive applications such as precision medicine. Here we present FrESCO: Framework for Exploring Scalable Computational Oncology, a modular deep-learning natural language processing (NLP) library for extracting pathology information from clinical text documents.
+
+
+## Quickstart Guide
+
+### Initial Setup
+Clone the repo
+```shell
+git clone https://code.ornl.gov/mossiac/fresco
+```
+Load the working branch and pull in the subrepos
+```shell
+cd fresco
+git checkout main
+```
+Setup the conda environment (the default name for the environment is "ms39", this can be edited in the ms39.yml file)
+```shell
+conda env create --file ms39.yml
+conda activate ms39
+```
+Install PyTorch 1.13.1. In Linux CUDA enabled setup, note that your specific cudatoolkit version requirements may vary,
+```shell
+conda install pytorch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1 pytorch-cuda=11.7 -c pytorch -c nvidia
+```
+otherwise, for CPU only,
+```shell
+conda install pytorch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1 cpuonly -c pytorch
+```
+Further PyTorch instructions may be found in the [PyTorch docs](https://pytorch.org/docs/stable/index.html).
+
+### Data Preparation
+The data used must be generated prior to inference. Add the path to existing data in the the 
+`data_path` argument. The required data files are:
+- `data_fold0.csv`: Pandas dataframe with columns:
+    - `X`: list of input values, of `int` type
+    - `task_n`: output for task `n`, a `string` type (these are the y-values)
+    - `split`: one of `train`, `test`, or `val`
+- `id2labels_fold0.json`: index to label dictionary mapping for each of the string representations of the outputs to an integer value, dict keys must match the y-values label
+- `word_embeds_fold0.npy`: word embedding matrix for the vocabulary, dimensions are `words x embedding_dim` 
+
+You will also need to set the `tasks`, these must correspond to the task columns in the `data_fold0.csv` file and keys in the `id2labels_fold0.json` dictionary.
+For example, in the P3B3 data, the task columns are `task_n, n = 1,2,3,4`. Whereas the imdb data has the `sentiment` task.
+If using any sort of class weighting scheme, the keyword `class_weights` must be either a pickle file or dictionary
+with keys corresponding to the task and value with a corresponding list, or numpy array, of weights for that task.  
+If the `class_weights` keyword is blank, corresponding to `None`, no class weighting scheme will be used during training nor inference.
+
+If working with hierarchical data, and the case-level context model is the desired output, then the dataframe in `data_fold0.csv` must contain an additional integer-valued column `group` where the values describe the hierarchy present within the data. For example, all rows where `group = 17` are associated for the purpose of training a case-level context model.
+
+### Model Training
+The `model_args.yaml` file controls the settings for model training. Edit this file as desired based on your requirements and desired outcome.
[…]
