### 2023-05-15 f4fb787
Add manuscript files
Adding manuscript files in moving to github.
