### 2023-05-17 DOE-NCI-MOSSAIC/FrESCO#1 @gabeerion
JOSS comments (Gabe)
Hi! I'm one of the reviewers on the paper submission to JOSS. I enjoyed reading the paper and agree this is a very useful area to develop software for! Thank you for contributing this exciting software. My review itself is in the joss-reviews issue and is mostly a list of checkboxes; I think the paper is well-written, the references are great, and the open-source nature of the software is well-established. I am posting here to discuss some specific areas in more depth than a check-box so we can address them and I can check them off my review! I will try to indicate with "ideally" comments/requests that are not mandatory but that from my perspective would greatly improve the value of the paper. Some of these comments may stem from a misunderstanding of the package or from something I missed in my own attempts to use it, and if so please correct me! Finally, I'm glad to break this out into separate specific issues in the repo if that's helpful.

# Scholarly effort
* The package does great along many of these axes. However, my understanding the main value proposition is to make it easy for non-NLP experts to abstract structured data from pathology reports. To that end, I think there's not enough documentation for a newcomer and at least some increase in documentation is necessary (see below).

# Installation
I was able to install the software by following the directions; however I don't currently see a reason the dependencies can't be handled automatically and/or that installation should require cloning the repo. In general a Python package submitted to JOSS should be pip-installable. If this isn't possible it would be good to know the reason for it. I also think if the package targets users with little programming knowledge it would be ideal for the pip install to provide an acutal command line utility (so that, for example, I could run something like `fresco --config cfg.yml --datadir path_to_dir --model_output path_to_output` from any directory, not just the repo).

# Functionality
I have encountered small issues so far that have prevented me from fully exploring the functionality of the code. They should be easily fixable. For example, on initially cloning the repo and running the scripts in `tests` I get an import error for the 'validate' module. I also get a KeyError on 'groups' on 'Step 2: Train a clc model' of the Readme.

# Documentation
Some of this is covered above, but I think there needs to be more documentation, which I will try to divide into essential areas and nice-to-haves
* Installation -- good, but would be nice to have automated dependency management (and see discussion of pip above).
* Example usage -- currently minimal. I think it is essential to have an example starting from a well-known dataset in a widely used format. For example, running through the IMDB example starting from a standard format of the IMDB dataset (ie from the Stanford website). Ideally, there would also be examples (possibly in notebook form) of each example dataset and model type, with a little more information on the datasets and model types included in the package. As someone new to this specific application area, I might want to know more about what the data usually look like, and as someone new to NLP, I might want to know more about which model to choose.
* Functionality documentation -- I think the package is intended to be usable by NLP practitioners as well, and they might want to make use of the Python functions. These appear to be well documented but ideally there would be a readthedocs type interface for them.
* Tests -- There are manual tests that I will evaluate more once they are running on my machine. Automated tests would be the ideal.

# Paper
Overall I think the paper is well-written. I have one comment, which is that ideally the statement of need would address why the models included in this software can't be run using their original implementations. My guess is this software aims to make them easier to use, and it would be great to provide a couple examples somewhere in the paper explaining why.

Thanks again for submitting this paper, and I would be very happy to chat more about any of these areas!
