### 2023-05-17 e17f6d9
update data_tests per PEP8
