+++ tests/model_args.yml
@@ -11,7 +11,7 @@ data_kwargs:
   tasks: ["task_1", "task_2", "task_3", "task_4"]
   # tasks: ['sentiment']  # ["task_1", "task_2", "task_3", "task_4"]
   fold_number: 0  # 0-based fold indexing
-  data_path: './data/P3B3/'
+  data_path: '../data/P3B3/'
   subset_proportion: 1.0
   add_noise_flag: True
   add_noise: 0.1
@@ -30,7 +30,7 @@ train_kwargs:
   max_epochs: 100
   patience: 5
   mixed_precision: False
-  class_weights: #  './weights/P3B3_weights.pickle' # path, dict, or blank
+  class_weights: #  './weights/random_weights.pickle' # path, dict, or blank
   # copied from data_kwargs
   # multilabel: False
   # random_seed: 1234
