+++ tests/profiling_test.py
@@ -1,221 +1,0 @@
-"""
-    Top-level model building script using independent modules.
-"""
-import argparse
-import json
-import os
-import random
-
-import torch
-import numpy as np
-
-from validate import exceptions
-from validate import validate_params
-from data_loaders import data_utils
-from abstention import abstention
-from keywords import keywords
-from models import mthisan, mtcnn
-from training import training
-from predict import predictions
-
-# needed for running on the vms
-# torch.multiprocessing.set_sharing_strategy('file_system')
-
-from torch.profiler import profile, record_function, ProfilerActivity
-
-def create_model(params, dw, device):
-    """Define model based on model_args.
-
-        Args:
-            params: dict of model_args file
-            dw: DataHandler class
-            device: torch.device, either 'cpu' or 'cuda'
-
-        Returns:
-            a model
-
-    """
-    if params.model_args['model_type'] == 'mthisan':
-        model = mthisan.MTHiSAN(dw.inference_data['word_embedding'],
-                                dw.num_classes,
-                                **params.model_args['MTHiSAN_kwargs'])
-
-    elif params.model_args['model_type'] == 'mtcnn':
-        model = mtcnn.MTCNN(dw.inference_data['word_embedding'],
-                            dw.num_classes,
-                            **params.model_args['MTCNN_kwargs'])
-    model.to(device)
-    if torch.cuda.device_count() > 1:
-        model = torch.nn.DataParallel(model)
-    return model
-
-
-def load_model(model_path, device, dw, valid_params):
-    """Load pretrained model from disk.
-
-        Args:
-            model_path: str, from command line args, points to saved model
-            device: torch.device, either cuda or cpu, set in calling function
-            dw: DataHandler class, for wrangling data, and checking metadata
-            valid_params: ValidateParams class, with model_args dict
-
-        We check if the supplied path is valid and if the packages match needed
-            to run the pretrained model.
-
-        Post-condition:
-            saved model loaded and set to eval() model
-
-    """
-    data_path = dw.model_args['data_kwargs']['data_path']
-    if os.path.exists(model_path):
-        print(f"Loading trained model from {model_path}")
-    else:
-        raise exceptions.ParamError(f'the model at {model_path} does not exist.')
-
-    if torch.cuda.is_available():
-        model_dict = torch.load(model_path)
-    else:
-        model_dict = torch.load(model_path, map_location=torch.device('cpu'))
-        model_dict = {k.replace('module.',''):v for k,v in model_dict.items()}
-
-    mismatches = [mod for mod in model_dict['metadata_package'].keys()
-        if dw.metadata['metadata'][mod] != model_dict['metadata_package'][mod]]
-
-    # check to see if the stored package matches the expected one
-    # what is this?
-    # package['match'] = len(mismatches)==0
-
-    if len(mismatches) > 0:
-        with open('metadata_package.json', 'w', encoding='utf-8') as jd:
-            json.dump(model_dict['metadata_package'], jd, indent=2)
-            raise exceptions.ParamError(f'the package(s) {", ".join(mismatches)} does not match ' +
-                                        f'the generated data in {data_path}.' +
-                                         '\nThe needed recreation info is in metadata_package.json')
-    # Load Model
-    if valid_params.model_args['model_type'] == 'mthisan':
-        model = mthisan.MTHiSAN(dw.inference_data['word_embedding'],
-                                dw.num_classes,
-                                **valid_params.model_args['MTHiSAN_kwargs'])
-
-    elif valid_params.model_args['model_type'] == 'mtcnn':
-        model = mtcnn.MTCNN(dw.inference_data['word_embedding'],
[…]
