+++ tests/perf_test.py
@@ -5,24 +5,17 @@ import argparse
 import json
 import os
 import random
+import sys
 
 import torch
-print(torch.__version__)
 import numpy as np
 
+sys.path.append("../")
 from validate import exceptions
 from validate import validate_params
 from data_loaders import data_utils
 from abstention import abstention
-from keywords import keywords
 from models import mthisan, mtcnn
-from training import training
-from predict import predictions
-
-# needed for running on the vms
-# torch.multiprocessing.set_sharing_strategy('file_system')
-
-torch.backends.cudnn.benchmark = True
 
 from torch.profiler import profile, record_function, ProfilerActivity
 
@@ -84,10 +77,6 @@ def load_model(model_path, device, dw, valid_params):
     mismatches = [mod for mod in model_dict['metadata_package'].keys()
         if dw.metadata['metadata'][mod] != model_dict['metadata_package'][mod]]
 
-    # check to see if the stored package matches the expected one
-    # what is this?
-    # package['match'] = len(mismatches)==0
-
     if len(mismatches) > 0:
         with open('metadata_package.json', 'w', encoding='utf-8') as jd:
             json.dump(model_dict['metadata_package'], jd, indent=2)
@@ -145,13 +134,11 @@ def main():
         os.makedirs('predictions')
     args = parser.parse_args()
 
-
     # 1. validate model/data args
     print("Validating kwargs in model_args.yml file")
-    cache_class = [False]
     data_source = 'pre-generated'
 
-    valid_params = validate_params.ValidateParams(cache_class, args, data_source=data_source)
+    valid_params = validate_params.ValidateParams(args, data_source=data_source)
 
     valid_params.check_data_train_args()
 
@@ -163,7 +150,7 @@ def main():
     if valid_params.model_args['abstain_kwargs']['abstain_flag']:
         valid_params.check_abstain_args()
 
-    valid_params.check_data_files(args.data_path)
+    valid_params.check_data_files()
 
     if valid_params.model_args['data_kwargs']['reproducible']:
         seed = valid_params.model_args['data_kwargs']['random_seed']
@@ -177,7 +164,7 @@ def main():
 
     # 2. load data
     print("Loading data and creating DataLoaders")
-    dw = data_utils.DataHandler(data_source, valid_params.model_args, cache_class)
+    dw = data_utils.DataHandler(data_source, valid_params.model_args)
     dw.load_folds()
 
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
@@ -196,62 +183,20 @@ def main():
                                              reproducible=valid_params.model_args['data_kwargs']['reproducible'],
                                              seed=seed)
 
-    if valid_params.model_args['train_kwargs']['keywords']:
-        kwds = keywords.Keywords(valid_params.model_args['data_kwargs']['tasks'],
-                                 dw.dict_maps['id2word'],
-                                 dw.dict_maps['id2label'],
-                                 device)
-        kwds.load_keyword_lists()
     # 3. create a model
     print("\nDefining a model")
     if len(args.model_path) == 0:
         model = create_model(valid_params, dw, device)
-        print("Device {0}".format(device))
+        print(f"Device {device}")
         sample = data_loaders['train']
-        sample
+        # sample
         for sample in data_loaders['train']:
             pass
         X = sample["X"].to(device)
-        with profile(activities=[ProfilerActivity.CPU, ProfilerActivity.CUDA],use_cuda=True, record_shapes=True) as prof:
+        with profile(activities=[ProfilerActivity.CPU, ProfilerActivity.CUDA], record_shapes=True) as prof:
             with record_function("model_inference"):
[…]
