### 2023-05-18 30c9126
absolute model path for train_model command line argument
