### 2023-05-18 bc22f63
Update data_setup script
