+++ fresco/predict/predictions.py
@@ -18,14 +18,15 @@ import torchmetrics
 
 
 class ScoreModel():
-    """Class to score and make predictions from a trained model.
+    """
+    Class to score and make predictions from a trained model.
 
-        Params:
-            model_args: dict with model_keyword args, checked by caller.
-            data_loaders: dict of PathReports class with split( train, test, val) as keys to
-                score and predict on
-            model: torch model fo predictiong and scoring
-            device: torch device, either cpu or gpu, set in caller
+    Args:
+        model_args (dict): Dictionary with model-specific keyword arguments, checked by the caller.
+        data_loaders (dict): Dictionary of PathReports class with split (train, test, val) as keys to
+            score and predict on.
+        model (torch model): Torch model for prediction and scoring.
+        device (torch.device): Torch device (either cpu or gpu) set in the caller.
 
     """
     def __init__(self, model_args, data_loaders, model, device, savepath=None, clc_flag=False):
@@ -87,17 +88,19 @@ class ScoreModel():
         self.metrics = {}
 
     def score(self, data_loaders=None, split=None, dac=None, training_phase=False):
-        """Score a dataset in a PathReport class.
+        """
+        Score a dataset in a PathReport class.
 
-            data_loaders: dict of PathReport class which provides X, y, and index of he entry in the
+        Args:
+            data_loaders (dict): Dictionary of PathReport class which provides X, y, and index of the entry in the
                 original dataframe, to tie back to the metadata.
-                The keys are 'train', 'test', and 'val' (if provided)
-            split: string, one of 'train', 'test', or 'val' if wanting to score only a
-                specific split
-            training_phase - bool: sets model to train or eval mode
-            dac - Abstention class: deep abstaining classifier class
+                The keys are 'train', 'test', and 'val' (if provided).
+            split (str): String, one of 'train', 'test', or 'val' if wanting to score only a specific split.
+            training_phase (bool): Sets model to train or eval mode.
+            dac (Abstention class): Deep abstaining classifier class.
 
         """
+
         if data_loaders is not None:
             print(("Error: score function requires a dict with keys 'train', 'test', or 'split'") +
                   (" and values PathReport classes."))
@@ -135,19 +138,21 @@ class ScoreModel():
         return None
 
     def _score(self, data_loader, split, training_phase=False, dac=None):
-        """Score data_loader for validation.
+        """
+        Score data_loader for validation.
 
-            Args:
-            data_loader - PathReports class
-            split: str of data split, eg, train, test, val
-            training_phase - bool: sets model to train or eval mode
-            dac - Abstention class: deep abstaining classifier class
+        Args:
+            data_loader (PathReports class): PathReports class.
+            split (str): String of data split, e.g., train, test, val.
+            training_phase (bool): Sets model to train or eval mode.
+            dac (Abstention class): Deep abstaining classifier class.
 
-            Post condition:
-                self.val_preds and self.val_trues updated.
-                model set to train or eval depending on training_phase variable
+        Post-condition:
+            self.val_preds and self.val_trues updated.
+            Model set to train or eval depending on training_phase variable.
 
         """
+
         losses = []
 
         if self.ntask:
@@ -179,7 +184,9 @@ class ScoreModel():
         self.scores[f'{split}_loss'] = np.mean(losses)
 
     def compute_scores(self, dac=None):  # , y_preds, y_trues, logits):
-        """Compute metrics to evaluate a model."""
+        """
+        Compute metrics to evaluate a model.
+        """
 
         metrics = {}
         # per task metrics
@@ -230,7 +237,9 @@ class ScoreModel():
         return metrics
 
     def save_scores(self):
-        """Write scores (metrics), per task, to disk."""
+        """
[…]
