+++ fresco/data_loaders/data_utils.py
@@ -22,7 +22,9 @@ from fresco.validate import exceptions
 
 
 class LabelDict(dict):
-    """Create dict subclass for correct behaviour when mapping task unks."""
+    """
+    Create dict subclass for correct behaviour when mapping task unks.
+    """
     def __init__(self, data, missing):
         super().__init__(data)
         self.data = data
@@ -34,22 +36,23 @@ class LabelDict(dict):
         return self.data[key]
 
     def __missing__(self, key):
-        """Assigns all missing keys to unknown in the label2id mapping."""
+        """
+        Assigns all missing keys to unknown in the label2id mapping.
+        """
         return self.data[self.missing]
 
 
 class DataHandler():
-    """Class for loading data.
-
-        Attributes:
-            data_source - str: defines how data was generated, needs to be
-                'pre-generated', 'pipeline', or 'official' (not implemented)
-            model_args - dict: keywords necessary to load data, build, and train a model_args
-
-        Note: the present implementation is only for 1 fold. We presently cannot
-        generate more than one fold of data.
-
     """
+    Class for loading data.
+
+    Attributes:
+        data_source (str): Defines how data was generated. Needs to be 'pre-generated', 'pipeline', or 'official' (not implemented).
+        model_args (dict): Keywords necessary to load data, build, and train a model_args.
+
+    Note: The present implementation is only for 1 fold. We presently cannot generate more than one fold of data.
+    """
+
     def __init__(self, data_source: str, model_args: dict, clc_flag: bool = False):
         self.data_source = data_source
         self.model_args = model_args
@@ -75,18 +78,18 @@ class DataHandler():
             self.grouped_cases = {}
 
     def load_folds(self, fold=None, subset_frac=None):
-        """Load data for each fold in the dataset.
+        """
+        Load data for each fold in the dataset.
 
-           Parameters:
-               fold: integer number of the fold to load
-               subset_frac: float val for proporiton to load
+        Parameters:
+            fold (int): Integer number of the fold to load.
+            subset_frac (float): Float value for proportion to load.
 
-           Pre-condition: self.__init__called and model_args is not None
+        Pre-condition: __init__ called and model_args is not None.
 
-           Post-condition: class attributes populated
+        Post-condition: Class attributes populated.
 
-           Case level context model will load fold 0 by default, see run_clc.py line 225.
-
+        Note: Case level context model will load fold 0 by default. See run_clc.py line 225.
         """
 
         data_loader = self.load_from_saved
@@ -128,15 +131,16 @@ class DataHandler():
                             t in self.tasks]
 
     def load_from_saved(self, fold: int, subset_frac: float = None) -> dict:
-        """Load data files.
-
-            Arguments: fold - int: fold number, should always be 0 for now
-
-            Post-condition:
-                Modifies self.splits in-place
-
-
         """
+        Load data files.
+
+        Arguments:
+            fold (int): Fold number. Should always be 0 for now.
+
+        Post-condition:
+            Modifies self.splits in-place.
+        """
+
         loaded_data = {'metadata': {}, "X": {}, "Y": {}}
         data_path = self.model_args['data_kwargs']['data_path']
 
@@ -192,7 +196,9 @@ class DataHandler():
         return loaded_data
 
     def get_vocab(self):
-        """Get the vocab from tokenized data."""
+        """
+        Get the vocab from tokenized data.
+        """
         
         embedding_dim = 300
 
@@ -212,17 +218,17 @@ class DataHandler():
         self.inference_data['word_embedding'] = rng.standard_normal(size=(vocab_len + 1, embedding_dim), dtype=np.float32) * 0.1
 
     def convert_y(self):
-        """Add task unknown labels to Y and map values to integers for inference.
-
-            Post-condition:
[…]

[…]
