### 2023-05-19 c052529
fixed up data utils
