+++ docs/build/html/_sources/index.rst.txt
@@ -1,0 +1,20 @@
+.. FrESCO documentation master file, created by
+   sphinx-quickstart on Thu May 18 17:04:31 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to FrESCO's documentation!
+==================================
+
+.. toctree::
+   :maxdepth: 4
+   :caption: Contents:
+
+   modules
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
