### 2023-05-19 DOE-NCI-MOSSAIC/FrESCO#1 @aspannaus
Hi @gabeerion,

Thanks for the comments and suggestions. We're presently working on getting the repo pip-installable, putting the documentation on readthedocs, and adding some notebook examples.

We can add some text to the manuscript describing why we cannot use the actual production code. The main development is in an air-gapped system with private health data, and we have restrictions on what can and cannot be moved outside. We believe that this repository is a reasonable compromise; if the required data files are supplied, the code runs similarly in both environments.

Regarding the issue you're seeing with training a clc model, I think the data path in the model_args file needs to point to the clc directory. The sample data files provided there has that missing key you mention above.

We've corrected the issue with the scripts in the tests folder as well.

We will finish internal testing with the reorganization to make the repo pip-installable and push the changes early next week.

Thanks again for the helpful comments and suggestions.
