### 2023-05-19 f384d15
update data setup script
