+++ data_setup.py
@@ -12,8 +12,8 @@ tqdm.pandas()
 
 def word2int(d, model):
     ints = [model.wv.key_to_index.get(d.lower()) for d in d.split(' ')]
-    X = [x for x in ints if x is not None]
-    return X
+    unk = len(model.wv.key_to_index)
+    return [x if x is not None else unk for x in ints]
 
 
 def main():
@@ -40,7 +40,7 @@ def main():
                                                     random_state=seed)
 
     x_val, x_test, y_val, y_test = train_test_split(x_tmp, y_tmp, test_size=test_split/(test_split + val_split),
-                                                    random_state=seed) 
+                                                    random_state=seed)
 
     train_df = pd.DataFrame(x_train, columns=['review', 'X', 'sentiment', 'split'])
     val_df = pd.DataFrame(x_val, columns=['review', 'X', 'sentiment', 'split'])
@@ -51,38 +51,33 @@ def main():
     test_df['split'] = 'test'
 
     train_df['sentiment'] = y_train
-    val_df['sentiment'] = y_val 
+    val_df['sentiment'] = y_val
     test_df['sentiment'] = y_test
 
+    train_data = pd.concat([train_df, val_df])
+
     print("Creating vocab and word embeddings")
-    model = gensim.models.word2vec.Word2Vec(vector_size=embed_dim, min_count=2, epochs=25, workers=8)
-    model.build_vocab(data)
-    model.train(data, total_examples=model.corpus_count, epochs=model.epochs)
-    
+    model = gensim.models.word2vec.Word2Vec(vector_size=embed_dim, min_count=2, epochs=2, workers=4)
+    model.build_vocab(train_data['review'].str.split())
+    model.train(train_data['review'].str.split(), total_examples=model.corpus_count, epochs=model.epochs)
+
     # tokenize data
     train_df['X'] = train_df['review'].progress_apply(lambda d: word2int(d, model))
     val_df['X'] = val_df['review'].progress_apply(lambda d: word2int(d, model))
     test_df['X'] = test_df['review'].progress_apply(lambda d: word2int(d, model))
-    
+
     # add <unk> token
     word_vecs = [model.wv.vectors[index] for index in model.wv.key_to_index.values()]
     rng = np.random.default_rng(seed)
     unk_embed = rng.normal(size=(1, embed_dim), scale=0.1)
     w2v = np.append(word_vecs, unk_embed, axis=0)
-    
-    # map words only in test to unk
-    train_data = pd.concat([train_df, val_df])
-    train_tok = set([tok for dat in train_data['X'] for tok in dat])
-    test_tok = set([tok for dat in test_df['X'] for tok in dat])
-    unks = list(set(test_tok).difference(train_tok))
-    w2v[unks] = unk_embed
-    
+
     id2word = {v: k for k, v in model.wv.key_to_index.items()}
     id2word[len(model.wv.key_to_index)] = "<unk>"
 
     print("Saving output files")
     df_out = pd.concat([train_df, val_df, test_df])
-    df_out.to_csv("./data/imdb/data_fold0.csv")
+    df_out.to_csv("./data/imdb/data_fold0.csv", index=False)
 
     labels = set(df['sentiment'])
     id2label = {'sentiment': {i: l for i, l in enumerate(labels)}}
