### 2023-05-21 #5345 @SimonBiggs
@jmschrei, just waiting on the following before I continue:

> We will finish internal testing with the reorganization to make the repo pip-installable and push the changes early next week.

https://github.com/DOE-NCI-MOSSAIC/FrESCO/issues/1#issuecomment-1554681881
