+++ build/lib/fresco/models/mthisan.py
@@ -1,0 +1,306 @@
+import math
+
+import torch
+from torch import nn
+import torch.nn.functional as F
+
+
+class MTHiSAN(nn.Module):
+    '''
+    Multitask hierarchical self-attention network for classifying cancer pathology reports.
+
+    Args:
+        embedding_matrix (numpy array): Numpy array of word embeddings.
+            Each row should represent a word embedding.
+            NOTE: The word index 0 is masked, so the first row is ignored.
+        num_classes (list[int]): Number of possible output classes for each task.
+        max_words_per_line (int): Number of words per line.
+            Used to split documents into smaller chunks.
+        max_lines (int): Maximum number of lines per document.
+            Additional lines beyond this limit are ignored.
+        att_dim_per_head (int, default: 50): Dimension size of output from each attention head.
+            Total output dimension is att_dim_per_head * att_heads.
+        att_heads (int, default: 8): Number of attention heads for multihead attention.
+        att_dropout (float, default: 0.1): Dropout rate for attention softmaxes and intermediate embeddings.
+        bag_of_embeddings (bool, default: False): Adds a parallel bag of embeddings layer.
+            Concats to the final document embedding.
+        embeddings_scale (float, default: 2.5): Scaling of word embeddings matrix columns.
+    '''
+
+    def __init__(self,
+                 embedding_matrix,
+                 num_classes,
+                 max_words_per_line,
+                 max_lines,
+                 att_dim_per_head=50,
+                 att_heads=8,
+                 att_dropout=0.1,
+                 bag_of_embeddings=False,
+                 embeddings_scale=2.5
+                 ):
+
+        super().__init__()
+        self.max_words_per_line = max_words_per_line
+        self.max_lines = max_lines
+        self.max_len = max_lines * max_words_per_line
+        self.att_dim_per_head = att_dim_per_head
+        self.att_heads = att_heads
+        self.att_dim_total = att_heads * att_dim_per_head
+
+        # normalize and initialize embeddings
+        embedding_matrix -= embedding_matrix.mean(axis=0)
+        embedding_matrix /= (embedding_matrix.std(axis=0, ddof=1) * embeddings_scale)
+
+        embedding_matrix[0] = 0
+        self.embedding = nn.Embedding.from_pretrained(
+                         torch.tensor(embedding_matrix, dtype=torch.float),
+                         freeze=False)
+        self.word_embed_drop = nn.Dropout(p=att_dropout)
+
+        # Q, K, V, and other layers for word-level self-attention
+        self.word_q = nn.Linear(embedding_matrix.shape[1], self.att_dim_total)
+        torch.nn.init.xavier_uniform_(self.word_q.weight)
+        self.word_q.bias.data.fill_(0.0)
+        self.word_k = nn.Linear(embedding_matrix.shape[1], self.att_dim_total)
+        torch.nn.init.xavier_uniform_(self.word_k.weight)
+        self.word_k.bias.data.fill_(0.0)
+        self.word_v = nn.Linear(embedding_matrix.shape[1], self.att_dim_total)
+        torch.nn.init.xavier_uniform_(self.word_v.weight)
+        self.word_v.bias.data.fill_(0.0)
+        self.word_att_drop = nn.Dropout(p=att_dropout)
+
+        # target vector and other layers for word-level target attention
+        self.word_target_drop = nn.Dropout(p=att_dropout)
+        self.word_target = nn.Linear(1, self.att_dim_total, bias=False)
+        torch.nn.init.uniform_(self.word_target.weight)
+        self.line_embed_drop = nn.Dropout(p=att_dropout)
+
+        # Q, K, V, and other layers for line-level self-attention
[…]
