### 2023-05-22 5f14b23
add data_setup script
