+++ README.md
@@ -18,41 +18,41 @@ git checkout main
 ```
 Setup the conda environment (the default name for the environment is "ms39", this can be edited in the ms39.yml file)
 ```shell
-conda env create --file ms39.yml
-conda activate ms39
+conda env create -n fresco python=3.9
+conda activate fresco
 ```
-Install PyTorch 1.13.1. In Linux CUDA enabled setup, note that your specific cudatoolkit version requirements may vary,
+Install the FrESCO library and dependencies
 ```shell
-conda install pytorch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1 pytorch-cuda=11.7 -c pytorch -c nvidia
+pip install fresco .
 ```
-otherwise, for CPU only,
-```shell
-conda install pytorch==1.13.1 torchvision==0.14.1 torchaudio==0.13.1 cpuonly -c pytorch
-```
-Further PyTorch instructions may be found in the [PyTorch docs](https://pytorch.org/docs/stable/index.html).
+
+For further PyTorch instructions or more details for pytorch, head over to the [PyTorch docs](https://pytorch.org/docs/stable/index.html).
 
 ### Data Preparation
 We have supplied three different datasets as examples, each must be
-unzipped before any model training via the `tar -xf dataset.tar.gz` command. The three datasets are:
+unzipped before any model training via the `tar -xf dataset.tar.gz` command from the `data` directory.
+The three datasets are:
   - imdb: binary sentiment classification with the imdb dataset,
   - P3B3: benchmark multi-task classification task, and
   - clc: case-level context multi-task classfication data.
 
-Add the path to the desired dataset in the the `data_path` argument in the `model_args.yml` file. The required data files are:
-- `data_fold0.csv`: Pandas dataframe with columns:
+Add the path to the desired dataset in the the `data_path` argument in the `configs/model_args.yml` file. The required data files are:
+    - `data_fold0.csv`: Pandas dataframe with columns:
     - `X`: list of input values, of `int` type
     - `task_n`: output for task `n`, a `string` type (these are the y-values)
     - `split`: one of `train`, `test`, or `val`
-- `id2labels_fold0.json`: index to label dictionary mapping for each of the string representations of the outputs to an integer value, dict keys must match the y-values label
-- `word_embeds_fold0.npy`: word embedding matrix for the vocabulary, dimensions are `words x embedding_dim`. If no word embedding exists, the model will use randomly generated embeddings.
+    - `id2labels_fold0.json`: index to label dictionary mapping for each of the string representations of the outputs to an integer value, dict keys must match the y-values label
+    - `word_embeds_fold0.npy`: word embedding matrix for the vocabulary, dimensions are `words x embedding_dim`. If no word embedding exists, the model will use randomly generated embeddings.
 
-You will also need to set the `tasks`, these must correspond to the task columns in the `data_fold0.csv` file and keys in the `id2labels_fold0.json` dictionary.
+You will also need to set the `tasks`, these must correspond to the task columns names in the `data_fold0.csv` file and keys in the `id2labels_fold0.json` dictionary.
 For example, in the P3B3 data, the task columns are `task_n, n = 1,2,3,4`. Whereas the imdb data has the `sentiment` task.
 If using any sort of class weighting scheme, the keyword `class_weights` must be either a pickle file or dictionary
 with keys corresponding to the task and value with a corresponding list, or numpy array, of weights for that task.  
 If the `class_weights` keyword is blank, corresponding to `None`, no class weighting scheme will be used during training nor inference.
 
-If working with hierarchical data, and the case-level context model is the desired output, then the dataframe in `data_fold0.csv` must contain an additional integer-valued column `group` where the values describe the hierarchy present within the data. For example, all rows where `group = 17` are associated for the purpose of training a case-level context model.
[…]
