+++ .readthedocs.yml
@@ -1,4 +1,4 @@
-version: 0.1 
+version: latest 
 
 build:
   os: "MacOS-13.3.1 (a)"
