+++ .readthedocs.yml
@@ -1,4 +1,4 @@
-version: "latest"
+version: 1
 
 build:
   os: "MacOS-13.3.1 (a)"
