### 2023-05-23 ecc3a0f
Update .readthedocs.yml
updated install
