### 2023-05-23 9d19e26
allow filename as cli and two-step clc args
