### 2023-05-23 1e0e903
add args files for each sample dataset
