### 2023-05-23 190ff9d
update clc model trainer
