+++ .readthedocs.yaml
@@ -1,7 +1,7 @@
 version: 2
 
 build:
-  os: "MacOS-13.3.1 (a)"
+  os: "ubuntu-20.04"
   tools:
     python: "3.10"
 
