+++ configs/clc_step2.yml
@@ -1,60 +1,30 @@
-save_name: 'clc_model'  # foldX.h5 will be appended at the end; leave blank to autopopulate with a timestamp
-task_unks:
-  task_1: 'n1'
-  task_2: 'b2'
-  task_3: 'b3'
-  task_4: 'c4'
-  sentiment: 'negative'
+save_name: 'clc_test'  # foldX.h5 will be appended at the end; leave blank to autopopulate with a timestamp
 
 # dataloader settings
 data_kwargs:
+  model_path: '../savedmodels/clc_model/model.h5'
   tasks: ["task_1", "task_2", "task_3", "task_4"]
-  fold_number: 0  # 0-based fold indexing
-  data_path: '../data/clc/'
-  subset_proportion: 1.0
-  add_noise_flag: True
-  add_noise: 0.1
-  mutual_info_filter: False
-  mutual_info_threshold: 0.0005
-  # copied to train_kwargs
+  exclude_single: False
+  shuffle_case_order: False
+  subset_proportion: 0.5
   reproducible: True
   random_seed: 42
-  batch_per_gpu: 128
-  doc_max_len: 3000
-  multilabel: False
 
 # general training args
-model_type: "mthisan" # mtcnn or mthisan
 train_kwargs:
+  batch_per_gpu: 128
   max_epochs: 100
   patience: 5
+  class_weights: #./weights/weights.pickle' # path, dict, or blank (None)
   mixed_precision: False
-  class_weights: #  './weights/P3B3_weights.pickle' # path, dict, or blank
-  # copied from data_kwargs
-  # multilabel: False
-  # random_seed: 1234
-  # batch_per_gpu: 128
-  # doc_max_len: 3000
 
-# mthisan model args
-MTHiSAN_kwargs:
-  max_words_per_line: 15
-  att_heads: 8
+# case level context model args
+model_kwargs:
   att_dim_per_head: 50
+  att_heads: 8
   att_dropout: 0.1
-  bag_of_embeddings: False
-  embeddings_scale: 2.5
+  forward_mask: True
 
-# mtcnn model args
-MTCNN_kwargs:
-  window_sizes: [3,4,5]
-  num_filters: [300,300,300]
-  dropout: 0.5
-  bag_of_embeddings: False
-  embeddings_scale: 20
-
-Transformers_kwargs:
-  
 # abstain args
 abstain_kwargs:
   abstain_flag: False
@@ -73,4 +43,3 @@ abstain_kwargs:
   ntask_alpha_scale: 0.8
   ntask_max_abs: 0.9
   ntask_min_acc: 0.975
-
