### 2023-05-23 fded1c4
added standalone command line interface.
