+++ build/lib/fresco/validate/validate_params.py
@@ -1,486 +1,0 @@
-"""
-
-Module for sanity checking parameters for data and model building.
-
-"""
-
-import datetime
-import json
-import math
-import os
-
-import yaml
-
-from fresco.validate import exceptions
-
-
-class ValidateParams():
-    """
-    Class to validate model-specific parameters for MOSSAIC models.
-
-    Args:
-        cli_args: argparse list of command line args.
-        data_source (str): Indicates where the data will come from. Should be one of:
-            - pre-generated: data_args.yml will indicate the source.
-
-    Post-condition: model_args dict loaded and sanity checked.
-    """
-
-    def __init__(self, cli_args,
-                 data_source: str = 'pre-generated',
-                 model_args: dict = None):
-
-        if model_args is None:
-            if len(cli_args.model_args) > 0:
-                mod_args_file = cli_args.model_args
-            else:
-                mod_args_file = 'model_args.yml'
-
-            if os.path.isfile(mod_args_file):
-                with open(mod_args_file, "r", encoding="utf-8") as f_in:
-                    self.model_args = yaml.safe_load(f_in)
-            else:
-                raise exceptions.ParamError("within FrESCO the " +
-                                            "model_args.yml file is needed to set " +
-                                            "the model arguments")
-        else:
-            self.model_args = model_args
-
-        if self.model_args['model_type'] not in ['mtcnn', 'mthisan']:
-            raise exceptions.ParamError("model type was not found " +
-                                        "to be 'mtcnn' or 'mthisan'. " +
-                                        "Currently these are the only expected options.")
-
-        if self.model_args['save_name'] == "":
-            self.save_name = f'model_{datetime.datetime.now().strftime("%Y%m%d%H%M%S")}'
-        else:
-            self.save_name = self.model_args['save_name']
-
-        self.save_name = os.path.join('savedmodels', self.save_name)
-
-        if not os.path.exists(os.path.dirname(self.save_name)):
-            print(f'savepath {os.path.dirname(self.save_name)} does not exist, creating it')
-            os.makedirs(os.path.dirname(self.save_name))
-
-        fold = self.model_args['data_kwargs']['fold_number']
-
-        if len(cli_args.data_path) > 0:
-            self.model_args['data_kwargs']['data_path'] = cli_args.data_path
-
-        if not os.path.exists(self.model_args['data_kwargs']['data_path']):
-            raise exceptions.ParamError("User provided data path does not exist.")
-
-        if isinstance(fold, int):
-            self.model_args['data_kwargs']['fold_number'] = fold
-        else:
-            raise exceptions.ParamError("Model building does not presently support > 1 fold.")
-
-        if (self.model_args['data_kwargs']['subset_proportion'] > 1 or
-                self.model_args['data_kwargs']['subset_proportion'] <= 0):
-            raise exceptions.ParamError("subset proportion must be float value between 0 and 1.")
-
-        if not isinstance(self.model_args['data_kwargs']['batch_per_gpu'], int):
-            raise exceptions.ParamError("Batch size must be an int value.")
-        if self.model_args['data_kwargs']['batch_per_gpu'] < 0 or \
-            self.model_args['data_kwargs']['batch_per_gpu'] > 2048:
-            raise exceptions.ParamError("Batch size must be an int value between 1 and 2048.")
-
-        if not isinstance(self.model_args['train_kwargs']['mixed_precision'], bool):
-            raise exceptions.ParamError("Mixed precision must be boolean.")
-
-    def hisan_arg_check(self):
-        """
-        Check and modify HiSAN specific args.
-
-        Parameters: none
-
[…]
