+++ build/lib/fresco/data_loaders/data_utils.py
@@ -1,625 +1,0 @@
-"""
-
-Module for loading pre-generated data.
-
-"""
-import json
-import math
-import os
-import pickle
-import random
-
-import torch
-
-import numpy as np
-import pandas as pd
-
-
-from torch.utils.data import DataLoader
-from torch.utils.data import Dataset
-
-from fresco.validate import exceptions
-
-
-class LabelDict(dict):
-    """
-    Create dict subclass for correct behaviour when mapping task unks.
-    """
-    def __init__(self, data, missing):
-        super().__init__(data)
-        self.data = data
-        self.missing = missing
-
-    def __getitem__(self, key):
-        if key not in self.data:
-            return self.__missing__(key)
-        return self.data[key]
-
-    def __missing__(self, key):
-        """
-        Assigns all missing keys to unknown in the label2id mapping.
-        """
-        return self.data[self.missing]
-
-
-class DataHandler():
-    """
-    Class for loading data.
-
-    Attributes:
-        data_source (str): Defines how data was generated. Needs to be 'pre-generated', 'pipeline', or 'official' (not implemented).
-        model_args (dict): Keywords necessary to load data, build, and train a model_args.
-
-    Note: The present implementation is only for 1 fold. We presently cannot generate more than one fold of data.
-    """
-
-    def __init__(self, data_source: str, model_args: dict, clc_flag: bool = False):
-        self.data_source = data_source
-        self.model_args = model_args
-
-        print(f"Loading data from {self.model_args['data_kwargs']['data_path']}")
-
-        self.inference_data = {}
-
-        self.dict_maps = {}
-
-        self.metadata = {'metadata': None, 'packages': None}
-
-        self.splits = []
-        self.num_classes = []
-        self.train_size = 0
-        self.val_size = 0
-        self.test_size = 0
-
-        self.tasks = self.model_args['data_kwargs']['tasks']
-        self.weights = None
-        if clc_flag:
-            self.clc_flag = True
-            self.grouped_cases = {}
-
-    def load_folds(self, fold=None, subset_frac=None):
-        """
-        Load data for each fold in the dataset.
-
-        Parameters:
-            fold (int): Integer number of the fold to load.
-            subset_frac (float): Float value for proportion to load.
-
-        Pre-condition: __init__ called and model_args is not None.
-
-        Post-condition: Class attributes populated.
-
-        Note: Case level context model will load fold 0 by default. See run_clc.py line 225.
-        """
-
-        data_loader = self.load_from_saved
-
-        if fold is None:
-            fold = self.model_args['data_kwargs']['fold_number']
-
-        if self.model_args['train_kwargs']['class_weights'] is not None:
-            self.load_weights(fold)
-
-        if subset_frac is None:
-            subset_frac = self.model_args['data_kwargs']['subset_proportion']
-        else:
-            subset_frac = subset_frac
-        loaded = data_loader(fold, subset_frac)
-
-        loaded['packages'] = {"mod_args": self.model_args,
-                              "py_packs": {"torch": str(torch.__version__),
-                                           "numpy": str(np.__version__),
-                                           "pandas": str(pd.__version__)},
-                              "fold": fold
-                             }  
-
-        self.inference_data['X'] = loaded['X']
-        self.inference_data['y'] = loaded['Y']
-        if loaded['we'] is None:
-            self.get_vocab() 
-        else:
-            self.inference_data['word_embedding'] = loaded['we']
-            self.dict_maps['id2word'] = loaded['id2word']
-
-        self.dict_maps['id2label'] = loaded['id2label']
-
-        self.metadata['metadata'] = loaded['metadata']
-
-        self.metadata['packages'] = loaded['packages']
-
[…]
