### 2023-05-24 73e268f
add step 2 of clc model
