+++ configs/clc_step2.yml
@@ -2,11 +2,11 @@ save_name: 'clc_test'  # foldX.h5 will be appended at the end; leave blank to au
 
 # dataloader settings
 data_kwargs:
-  model_path: '../savedmodels/clc_model/model.h5'
+  model_path: './savedmodels/clc_model/clc_model.h5'
   tasks: ["task_1", "task_2", "task_3", "task_4"]
   exclude_single: False
   shuffle_case_order: False
-  subset_proportion: 0.5
+  subset_proportion: 1.0
   reproducible: True
   random_seed: 42
 
