### 2023-05-24 DOE-NCI-MOSSAIC/FrESCO#1 @aspannaus
Hey @gabeerion and @[SimonBiggs](https://github.com/SimonBiggs),

We've made our edits based on @gabeerion's helpful comments and suggestions. Please continue with the review and let us know what you all think. Thanks!
