### 2023-05-24 9c19853
Move data_setup to scripts dir
