### 2023-05-25 #5345 @aspannaus
Just posting here for completeness, we updated the repository based on @gabeerion comments and suggestions.

[https://github.com/DOE-NCI-MOSSAIC/FrESCO/issues/1#issuecomment-1561201091](https://github.com/DOE-NCI-MOSSAIC/FrESCO/issues/1#issuecomment-1554681881)
