### 2023-06-05 9ffc55c
Unix compatibility of docs
OSX and Linux are case sensitive with the `cd` command
