### 2023-06-06 27645da
Update README.md
Add `pypi` install.
