+++ docs/source/index.rst
@@ -3,13 +3,269 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to FrESCO's documentation!
-==================================
+FrESCO: Framework for Exploring Scalable Computational Oncology
+===============================================================
 
 .. image:: extra/logo.png
   :width: 400
   :alt: FrESCO logo
 
+
+Motivation
+----------
+
+The National Cancer Institute (NCI) monitors population level cancer
+trends as part of its Surveillance, Epidemiology, and End Results (SEER)
+program. This program consists of state or regional level cancer
+registries which collect, analyze, and annotate cancer pathology
+reports. From these annotated pathology reports, each individual
+registry aggregates cancer phenotype information and summary statistics
+about cancer prevalence to facilitate population level monitoring of
+cancer incidence. Extracting cancer phenotype from these reports is a
+labor intensive task, requiring specialized knowledge about the reports
+and cancer. Automating this information extraction process from cancer
+pathology reports has the potential to improve not only the quality of
+the data by extracting information in a consistent manner across
+registries, but to improve the quality of patient outcomes by reducing
+the time to assimilate new data and enabling time-sensitive applications
+such as precision medicine. Here we present FrESCO: Framework for
+Exploring Scalable Computational Oncology, a modular deep-learning
+natural language processing (NLP) library for extracting pathology
+information from clinical text documents.
+
+Documentation is available over at `Read the
+Docs <https://doe-nci-fresco.readthedocs.io/en/latest/>`__.
+
+Quickstart Guide
+----------------
+
+Install via pip
+~~~~~~~~~~~~~~~
+
+Install from `pypi <https://pypi.org/project/nci-fresco/>`__ using
+
+.. code:: shell
+
+   pip install nci-fresco
+
+Install from source
+~~~~~~~~~~~~~~~~~~~
+
+Clone the repo
+
+.. code:: shell
+
+   git clone https://github.com/DOE-NCI-MOSSAIC/FrESCO.git
+
+Load the working branch and pull in the subrepos
+
+.. code:: shell
+
+   cd FrESCO
+   git checkout main
+
+Setup the conda environment (the default name for the environment is
+“ms39”, this can be edited in the ms39.yml file)
+
+.. code:: shell
+
+   conda env create -n fresco python=3.9
+   conda activate fresco
+
+Install the FrESCO library and dependencies
+
+.. code:: shell
+
+   pip install .
+
+For further PyTorch instructions or more details for pytorch, head over
+to the `PyTorch docs <https://pytorch.org/docs/stable/index.html>`__.
+
+Notebooks and Examples
+~~~~~~~~~~~~~~~~~~~~~~
+
+We have supplied example notebooks for each of the sample datasets
+contained in this repository showing model training on each dataset. We
+have also supplied ``model_args`` files for each of the datasets
+contained within the repo to speedup the time to ge up and running with
+the codebase.
+
+Data Preparation
+~~~~~~~~~~~~~~~~
+
+We have supplied three different datasets as examples, each must be
+unzipped before any model training via the ``tar -xf dataset.tar.gz``
+command from the ``data`` directory. The three datasets are: - imdb:
+binary sentiment classification with the imdb dataset, - P3B3: benchmark
+multi-task classification task, and - clc: case-level context multi-task
+classfication data.
+
+We have prepared ``model_args`` files for each dataset within the
+``configs/`` directory. To run with your own data, the following
+instructions explain the requirements for the training data.
+
+Add the path to the desired dataset in the the ``data_path`` argument in
+the ``configs/model_args.yml`` file. The required data files are: -
+``data_fold0.csv``: Pandas dataframe with columns: - ``X``: list of
[…]
