### 2023-06-26 4c23915
Update README and add notebook for sklearn-style model
