+++ README.md
@@ -12,6 +12,7 @@ Install from [pypi](https://pypi.org/project/nci-fresco/) using
 ```shell
 pip install nci-fresco
 ```
+As an alternative, you may download the source code.
 
 ### Install from source
 Clone the repo
@@ -40,6 +41,8 @@ For further PyTorch instructions or more details for pytorch, head over to the [
 We have supplied example notebooks for each of the sample datasets contained in this repository showing 
 model training on each dataset. We have also supplied `model_args` files for each of the 
 datasets contained within the repo to speedup the time to ge up and running with the codebase. 
+
+
 
 ### Data Preparation
 We have supplied three different datasets as examples, each must be
@@ -130,6 +133,21 @@ Training a model with the supplied default args, we see convergence within 50-60
 in excess of 0.90 accuracy across all tasks for the P3B3 dataset within 60 epochs or so. **NOTE:** P3B3 has a known issue training with mixed
 precision and with `DAC` and `NTask` enabled. Ensure these keywords are all `False` for all runs with the P3B3 dataset.
 
+
+### Bring Your Own (Torch) Model
+
+To try out your favorite torch model with our training and evaluation methods, your torch model must subclass the `torch.nn.Module` 
+and provide a `forward` call within the model class. The `forward` method is expected to take a pytorch `torch.utils.data.DataLoader` class
+batched input, a map-style dataset, see the [docs](https://pytorch.org/docs/stable/data.html#map-style-datasets) describing this class.
+The `forward` method is expected to return 'logits', the unnormalized outputs fro mthe network that have not been passed through a
+function mapping the logits the interval [0, 1].
+
+The machinery provied here is designed to work with nequrla network models, and is overkill for a traditional sklearn type
+[model](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html), as most of the necessary
+functions to train and evaluate a model are provided within sklearn. If you wish to try our sample pre-processing 
+steps with your favorite sklearn model, an example is provided within the notebooks directory. Using the
+generated word embeddings and tokenized documents generated by our preprocessing script are incompatibe with
+the sklearn library, which expects feature vectors as input, not higher dimensional inputs. 
 
 ### Contributing
 
