### 2023-06-26 #5345 @gabeerion
Hi, so sorry for the delay! I appreciate the ping and will work on this again tomorrow.

Get Outlook for Android<https://aka.ms/AAb9ysg>
