### 2023-07-06 #5345 @gabeerion
Thanks for the reminder - I am working through it but a bit slowly due to other work. I think I can solidly say I'll be done within a week.

Get Outlook for Android<https://aka.ms/AAb9ysg>
