### 2023-07-18 DOE-NCI-MOSSAIC/FrESCO#1 @gabeerion
Hi all, thank you for your hard work on the repo! I have updated my review in the joss-reviews thread and only have one or two comments left. I think the last two points I have are regarding functionality documentation and example usage:
1 - The notebooks are very helpful, I think the last example usage that would be good is showing how to go start with a trained model and create/save predictions on a new unlabeled dataset.
2 - For documentation, brief explanations of all the config.yml parameters seems important to me, ideally in the readthedocs somewhere.

If these can be completed I think I'll be able to finalize the review in short order. Thank you!
