### 2023-07-24 5356a3c
Update script for predictions from trained model
