### 2023-07-25 9fcd2c2
Update use_model.py
add comment in use_model for expected use
