### 2023-07-31 DOE-NCI-MOSSAIC/FrESCO#1 @aspannaus
Hey @gabeerion,

Thanks for the comments and feedback. We've updated the repository with an additional example notebook and description of the args in the config.yml file. Let us know if you see anything else.

Thanks!
