### 2023-08-03 #5345 @anupamajha1
@aspannaus Can you address the following points so that I can complete the corresponding review sections?

1. Can you add who the target audience of FrESCO is?  I can then check off the "A statement of need" check box.

2. What does MIMIC stand for in the "State of the field" section of the paper?

3. There was the following warning when decompressing the inference.tar.gz file:
"tar: Ignoring unknown extended header keyword `LIBARCHIVE.xattr.com.apple.quarantine'"

4. Could not run one of the provided models using the commands from the README file. Added details in this issue (https://github.com/DOE-NCI-MOSSAIC/FrESCO/issues/32),

4. The data preparation and model training sections of the READme are extremely complicated, some complexity can be simplified. For example, the model training section states that to run the model, run this command, "python train_model.py -m ie -args ../configs/model_args.yml." But this command cannot be run as is, the model_args.yml needs to be replaced with one of the actual yml files from the config directory.

5. Furthermore, the code in the scripts/ directory has very few comments. Please add informative comments to aid understanding of your code flow for a layperson.
