### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#32 @aspannaus
Hi @anupamajha1,

Thanks for looking at this. I think it is related to a comment from the other JOSS reviewer, and we edited the config file, removing unnecessary args. Which branch are you on, and when did you pull it? I pulled a fresh version of `main` and step 1 of clc runs as expected. If you can provide these details, we should be able to get things straightened out in short order.

Thanks
