### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#32 @anupamajha1
Hi @aspannaus, I am on the main branch and I ran this on code I pulled today morning (Aug 2, 2023). However, after your comment, I pulled your latest changes from the main branch (Aug 2, 2023, 7:20 PM PST) and reinstalled everything and ran the command again but I still get the same error as before (see below).

```
python train_model.py -m ie -args ../configs/clc_step1.yml

Process started
Wed Aug  2 19:21:52 2023

Running information extraction task
Validating kwargs in model_args.yml file

Received:  ['add_noise', 'add_noise_flag', 'batch_per_gpu', 'data_path', 'doc_max_len', 'fold_number', 'multilabel', 'random_seed', 'reproducible', 'subset_proportion', 'tasks']
Expected:  ['add_noise', 'add_noise_flag', 'batch_per_gpu', 'data_path', 'doc_max_len', 'fold_number', 'multilabel', 'mutual_info_filter', 'mutual_info_threshold', 'random_seed', 'reproducible', 'subset_proportion', 'tasks']
Traceback (most recent call last):
  File "/net/noble/vol3/user/anupamaj/proj/2023_openjournal_review/FrESCO/scripts/train_model.py", line 64, in <module>
    main()
  File "/net/noble/vol3/user/anupamaj/proj/2023_openjournal_review/FrESCO/scripts/train_model.py", line 51, in main
    run_ie.run_ie(args)
  File "/net/noble/vol1/home/anupamaj/miniconda3/fresco/lib/python3.9/site-packages/fresco/run_ie.py", line 141, in run_ie
    valid_params.check_data_train_args()
  File "/net/noble/vol1/home/anupamaj/miniconda3/fresco/lib/python3.9/site-packages/fresco/validate/validate_params.py", line 175, in check_data_train_args
    raise exceptions.ParamError((f"model args {kwrd} does not have " +
fresco.validate.exceptions.ParamError: model args data_kwargs does not have the expected variables
```
