### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#32 @anupamajha1
I deleted everything and reinstalled the repository from scratch again at 7:46 PM PST, Aug 2, 2023, and I still get the same error. I will wait for the resolution of this issue for my review.
