+++ fresco/data_loaders/data_utils.py
@@ -77,7 +77,7 @@ class DataHandler():
             self.clc_flag = True
             self.grouped_cases = {}
 
-    def load_folds(self, fold: int, subset_frac: float = None):
+    def load_folds(self, fold: int = 0, subset_frac: float = None):
         """
         Load data for each fold in the dataset.
 
@@ -112,12 +112,12 @@ class DataHandler():
                                            "numpy": str(np.__version__),
                                            "pandas": str(pd.__version__)},
                               "fold": fold
-                             }  
+                             }
 
         self.inference_data['X'] = loaded['X']
         self.inference_data['y'] = loaded['Y']
         if loaded['we'] is None:
-            self.get_vocab() 
+            self.get_vocab()
         else:
             self.inference_data['word_embedding'] = loaded['we']
             self.dict_maps['id2word'] = loaded['id2word']
@@ -171,7 +171,7 @@ class DataHandler():
             loaded_data['metadata'][split] = df[df['split'] == split][[v for v in df.columns
                                                              if v not in ['X', *loaded_data['id2label'].keys()]]]
             data_tasks = set([v for v in df.columns if v in loaded_data['id2label'].keys()])
-            if len(data_tasks) > 0: 
+            if len(data_tasks) > 0:
                 loaded_data['Y'][split] = df[df['split'] == split][sorted(loaded_data['id2label'].keys())]
 
         if subset_frac < 1:
@@ -203,14 +203,14 @@ class DataHandler():
         """
         Get the vocab from tokenized data.
         """
-        
+
         embedding_dim = 300
 
         X = self.inference_data['X']['train']
-        
+
         if 'val' in self.inference_data['X'].keys():
-            X = pd.concat([X, self.inference_data['X']['val']])        
-        
+            X = pd.concat([X, self.inference_data['X']['val']])
+
         # find all unique tokens
         s1 = set(X.iloc[0]).union(X.iloc[1])
 
@@ -261,7 +261,7 @@ class DataHandler():
         Loads class weights from pickle file or dict in model_args file.
 
         Args:
-            fold (int): Data fold to be loaded 
+            fold (int): Data fold to be loaded
 
         """
 
@@ -275,7 +275,7 @@ class DataHandler():
             tmp = json.load(f)
         id2label = {task: {int(k): str(v) for k, v in labels.items()}
                            for task, labels in tmp.items()}
-		
+
         if isinstance(self.model_args['train_kwargs']['class_weights'], dict):
             self.weights = self.model_args['train_kwargs']['class_weights']
 
@@ -425,7 +425,7 @@ class DataHandler():
                                                 clc_args['train_kwargs']['batch_per_gpu'],
                                                 shuffle=False) for split in self.splits}
 
-    def inference_loader(self, reproducible: bool = True, 
+    def inference_loader(self, reproducible: bool = True,
                          seed: int = None, batch_size: int = 128) -> dict:
         """
         Create torch DataLoader class for inference from a trained model.
