### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#32 @aspannaus
Hi @anupamajha1,

I've figured out the issue; the code on pypi is slightly older than that in the repo. If you run `pip install -e .` it runs as expected. The person on the team who set up the pypi account is out right now, and we'll update that as well once they're back in the office.

Thanks
