+++ scripts/train_model.py
@@ -3,6 +3,11 @@
 """
     Top level module to create information extraction or case-level context models.
 
+    Typical call is $python train_model.py -m ie -args ../configs/model_args.yml
+    This will train a model on the P3B3 dataset. The model_args.yml
+    file may be swapped for any of the provided yml files, or 
+    a custom config file. 
+
 """
 
 import argparse
