### 2023-08-03 #5345 @aspannaus
Hi @anupamajha1,

Thanks for the comments on out repo. Our responses follow:

1. We added text in the paper at the end of the first paragraph.
2. This has been defined in the second paragraph.
3. This has been fixed.
4. This has been addressed and will be pushed to pypi next week. I'll message here when it is updated on pypi.
5. The generic `model_args.yml` file will train a model on the `P3B3` dataset. This has been noted in the `README.md`. We also made some changes to the `README` to reduce the complexity and improve the organization. If there are more specific ideas that you have, we're happy to iterate.
6. We have added more comments on the expected use of the scripts in the `scripts/` folder and more inline comments.

Thanks for the comments and helpful suggestions.
