### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#32 @anupamajha1
Hi @aspannaus, Thanks for addressing the issue, I can run train_model.py if I pip install with -e option. Please don't forget to update the pypi version.
