### 2023-08-03 DOE-NCI-MOSSAIC/FrESCO#34 @anupamajha1
Macro precision and recall are sometimes > 1 for P3B3 model test set
Hi,

I ran the following command:

`python train_model.py -m ie -args ../configs/model_args.yml`

This model trains FrESCO for P3B3 dataset. While looking at the metrics produced by the model on the test set, I noticed >1 macro precision at one place (see attached metric file) and >1 macro recall at another (attached metric file).  Please check if this is the expected behavior and if not why the metrics are breaking.
[Fresco_P3B3_model.tar.gz](https://github.com/DOE-NCI-MOSSAIC/FrESCO/files/12254645/Fresco_P3B3_model.tar.gz)
[model_metrics_20230803125159.csv](https://github.com/DOE-NCI-MOSSAIC/FrESCO/files/12254648/model_metrics_20230803125159.csv)
[model_metrics_20230803125216.csv](https://github.com/DOE-NCI-MOSSAIC/FrESCO/files/12254650/model_metrics_20230803125216.csv)
