### 2023-08-03 #5345 @anupamajha1
@aspannaus Thanks for addressing my comments. I was able to run the code after using `pip install -e . ` command. However, I noticed that FrESCO sometimes produces >1 macro precision and recall and I have started a new issue to let the authors address if this is okay and if not, what is breaking the code (https://github.com/DOE-NCI-MOSSAIC/FrESCO/issues/34)
