+++ fresco/predict/predictions.py
@@ -226,13 +226,13 @@ class ScoreModel():
             num_classes = scores.shape[1]
 
             # macro hack
-            _cat_trues_preds = torch.cat((_trues, _preds), dim=0)
-            scale_classes = num_classes / len(torch.unique(_cat_trues_preds))
+            # _cat_trues_preds = torch.cat((_trues, _preds), dim=0)
+            # scale_classes = num_classes / len(torch.unique(_cat_trues_preds))
 
             f1_micro = torchmetrics.F1Score(task='multiclass', num_classes=num_classes, average='micro').to(self.device)
             metrics[f'{task}_micro'] = f1_micro(_preds, _trues)
             f1_macro = torchmetrics.F1Score(task='multiclass', num_classes=num_classes, average='macro').to(self.device)
-            metrics[f'{task}_macro'] = f1_macro(_preds, _trues) * scale_classes
+            metrics[f'{task}_macro'] = f1_macro(_preds, _trues)  # * scale_classes
 
             accuracy = torchmetrics.Accuracy(task='multiclass', num_classes=num_classes).to(self.device)
             metrics[f'{task}_accuracy'] = accuracy(_preds, _trues)
@@ -240,12 +240,12 @@ class ScoreModel():
             precision_micro = torchmetrics.Precision(task='multiclass', num_classes=num_classes, average='micro').to(self.device)
             metrics[f'{task}_precision_micro'] = precision_micro(_preds, _trues)
             precision_macro = torchmetrics.Precision(task='multiclass', num_classes=num_classes, average='macro').to(self.device)
-            metrics[f'{task}_precision_macro'] = precision_macro(_preds, _trues) * scale_classes
+            metrics[f'{task}_precision_macro'] = precision_macro(_preds, _trues)  # * scale_classes
 
             recall_micro = torchmetrics.Recall(task='multiclass', num_classes=num_classes, average='micro').to(self.device)
             metrics[f'{task}_recall_micro'] = recall_micro(_preds, _trues)
             recall_macro = torchmetrics.Recall(task='multiclass', num_classes=num_classes, average='macro').to(self.device)
-            metrics[f'{task}_recall_macro'] = recall_macro(_preds, _trues) * scale_classes
+            metrics[f'{task}_recall_macro'] = recall_macro(_preds, _trues)  # * scale_classes
 
 
         if self.abstain:
