+++ fresco/training/training.py
@@ -294,12 +294,17 @@ class ModelTrainer():
 
             if val_scores['val_loss'][0] < best_loss:
                 best_loss = val_scores['val_loss'][0]
-                torch.save(self.model.state_dict(), self.savename)
+                torch.save({'epoch': epoch,
+                            'model_state_dict': self.model.state_dict(), 
+                            'opt_state_dict': self.opt.state_dict(),
+                            'val_loss': best_loss
+                            }, self.savename)
 
             if stop:
                 print(f"saving to {self.savename}", flush=True)
                 # loading weights of best model
-                self.model.state_dict(torch.load(self.savename))
+                checkpoint = torch.load(self.savename)
+                self.model.state_dict(checkpoint['model_state_dict'])
                 scores = f"epoch_{epoch}_scores_fold{self.fold}_{datetime.datetime.now().strftime('%Y%m%d%H%M%S')}.pkl"
                 with open(self.savepath + scores, "wb") as f_out:
                     pickle.dump(all_scores, f_out, pickle.HIGHEST_PROTOCOL)
@@ -307,7 +312,8 @@ class ModelTrainer():
         if epoch + 1 == self.epochs:
             print('\nModel training hit max epochs, not converged')
             # loading weights of best model
-            self.model.state_dict(torch.load(self.savename))
+            checkpoint = torch.load(self.savename)
+            self.model.state_dict(checkpoint['model_state_dict'])
             print(f"saving to {self.savename}", flush=True)
             scores = f"epoch_{epoch}_scores_fold{self.fold}_{datetime.datetime.now().strftime('%Y%m%d%H%M%S')}.pkl"
             with open(self.savepath + scores, "wb") as f_out:
