### 2023-08-07 e1f9606
Update random embedding matrix generation
