+++ fresco/data_loaders/data_utils.py
@@ -21,6 +21,16 @@ from torch.utils.data import Dataset
 from fresco.validate import exceptions
 
 
+def word2int(tok, vocab):
+    """Map words to tokens for random embeddings.
+
+        If a word doesn't exist in the training/val split, it is mapped
+        to 'unk', the unknown token.
+
+    """
+    unk = len(vocab)
+    return [x if x is not None else unk for x in tok]
+
 class LabelDict(dict):
     """
     Create dict subclass for correct behaviour when mapping task unks.
@@ -201,7 +211,7 @@ class DataHandler():
 
     def get_vocab(self):
         """
-        Get the vocab from tokenized data.
+        Get the vocab and word embeddings from tokenized data.
         """
 
         embedding_dim = 300
@@ -212,14 +222,20 @@ class DataHandler():
             X = pd.concat([X, self.inference_data['X']['val']])
 
         # find all unique tokens
-        s1 = set(X.iloc[0]).union(X.iloc[1])
-
+        s1 = set(X.iloc[0])
         for x in X:
             s1 = set(x).union(s1)
 
         vocab_len = np.max(list(s1))
+        unk = vocab_len + 1 
+        self.inference_data['X']['test'].apply(lambda d: word2int(d, s1))
         rng = np.random.default_rng(self.model_args['train_kwargs']['random_seed'])
-        self.inference_data['word_embedding'] = rng.standard_normal(size=(vocab_len + 1, embedding_dim), dtype=np.float32) * 0.1
+        unk_embed = rng.normal(size=(1, embedding_dim), scale=0.1)
+        random_embeds = rng.standard_normal(size=(vocab_len, embedding_dim), dtype=np.float32) * 0.1
+        self.inference_data['word_embedding'] = np.concatenate((np.zeros(shape=(1, embedding_dim)),
+                                                                random_embeds,
+                                                                unk_embed),
+                                                                axis=0)
 
     def convert_y(self):
         """
