### 2023-08-08 DOE-NCI-MOSSAIC/FrESCO#1 @gabeerion
Hi, thanks for these edits! I think they're great. I'm going to recommend two changes but I don't think either of them are essential to finish my review so I will mark it complete.

1 - It would be good for the example notebook to mention that it is essentially  a breakdown of how `use_model.py` works -- it took me til halfway through or so to realize

2 - The explanations of the args are very helpful! For some reason they're formatted nicely on Github but the formatting breaks on readthedocs. It'd be nice to fix -- but again, not worth delaying my review completion over.

Thanks for sharing the great work!
