### 2023-08-08 #5345 @aspannaus
Hi @anupamajha1,

We've looked into the issue you describe, and while we weren't able to reproduce it, we found an issue with the way that the macro scores were being computed within torchmetrics, see: https://github.com/Lightning-AI/torchmetrics/issues/1664 and https://github.com/Lightning-AI/torchmetrics/pull/1821.

We've modified the code to call the sklearn f1 score function, tested it, and everything is working as expected. The latest version has been updated on pypi as well. Let us know if there's anything else you need from us.

Thanks!
