### 2023-08-08 DOE-NCI-MOSSAIC/FrESCO#1 @aspannaus
Hi @gabeerion,

Thanks for the helpful comments and feedback on the repo. We went ahead and made a few edits based on your comments above; it makes everything flow a bit more smoothly.

Thanks again!
