### 2023-08-09 dba6146
Update training.py
remove unused functions
