### 2023-08-11 1a8b4de
capture return val from series label mapping
