+++ paper/shared/refs.bib
@@ -49,7 +49,7 @@
 	publisher = {Oxford University Press},
 	title = {Automatic extraction of cancer registry reportable information from free-text pathology reports using multitask convolutional neural networks},
 	volume = {27},
-	doi{10.1093/jamia/ocz153},
+	doi={10.1093/jamia/ocz153},
 	year = {2020}}
 
 @article{GAO2019101726,
