+++ setup.py
@@ -2,19 +2,20 @@ from setuptools import setup, find_packages
 
 setup(
     name='fresco',
-    version='0.1.0',    
+    version='0.2.0',    
     description='',
     url='https://github.com/DOE-NCI-MOSSAIC/FrESCO',
     author='Adam Spannaus',
     author_email='spannausat@ornl.gov',
     license='MIT',
     packages=find_packages(include=['fresco', 'fresco.*']),
-    install_requires=[ 'pandas',                     
+    install_requires=['pandas',                     
                       'scikit-learn',                     
                       'pyyaml',                     
                       'gensim',                     
                       'numpy',                     
-                      'tqdm',       
+                      'tqdm', 
+                      'polars'      
                       ],
 
     classifiers=[
