+++ fresco/data_loaders/data_utils.py
@@ -13,6 +13,7 @@ import torch
 
 import numpy as np
 import pandas as pd
+import polars as pl
 
 
 from torch.utils.data import DataLoader
@@ -357,11 +358,8 @@ class DataHandler():
             worker = None
             gen = None
 
-        if torch.cuda.is_available():
-            pin_mem = True
-        else:
-            pin_mem = False
-
+        pin_mem = bool(torch.cuda.is_available())
+        
         loaders = {}
 
         vocab_size = self.inference_data['word_embedding'].shape[0]
@@ -423,23 +421,33 @@ class DataHandler():
 
         return loaders
 
-    def make_grouped_cases(self, doc_embeds,clc_args, device):
-        """
-        Created GroupedCases class for torch DataLoaders.
-        """
+    def make_grouped_cases(self, doc_embeds, clc_args, device, reproducible:bool = True, seed: int = None):
+        """Created GroupedCases class for torch DataLoaders."""
 
-        datasets = {split: GroupedCases(doc_embeds[split],
-                                        self.inference_data['y'][split],
+        if reproducible:
+            gen = torch.Generator()
+            gen.manual_seed(seed)
+            worker = self.seed_worker
+        else:
+            worker = None
+            gen = None
+        
+        datasets = {split: GroupedCases(doc_embeds[split]['X'],
+                                        doc_embeds[split]['y'],
+                                        doc_embeds[split]['index'],
                                         self.model_args['data_kwargs']['tasks'],
                                         self.metadata['metadata'][split],
                                         device,
                                         exclude_single=clc_args['data_kwargs']['exclude_single'],
-                                        shuffle_data_order=clc_args['data_kwargs']['shuffle_case_order'])
-                                        for split in self.splits}
+                                        shuffle_case_order=clc_args['data_kwargs']['shuffle_case_order'],
+                                        split_by_tumor_id=clc_args['data_kwargs']['split_by_tumorid']
+                                        ) for split in self.splits}
 
         self.grouped_cases = {split: DataLoader(datasets[split],
-                                                clc_args['train_kwargs']['batch_per_gpu'],
-                                                shuffle=False) for split in self.splits}
+                                                batch_size=clc_args['train_kwargs']['batch_per_gpu'],
+                                                shuffle=False,
+                                                worker_init_fn=worker,
+                                                generator=gen) for split in self.splits}
 
     def inference_loader(self, reproducible: bool = True,
                          seed: int = None, batch_size: int = 128) -> dict:
@@ -598,104 +606,110 @@ class PathReports(Dataset):
         return sample
 
 class GroupedCases(Dataset):
-    """Class for grouping cases for clc."""
+    """Create grouped cases for torch DataLoaders.
 
+        args:
+            doc_embeds - document embeddings from trained model as np.ndarray
+            Y - dict of integer Y values, keys are the splits
+            tasks - list of tasks
+            metadata - dict of model metadata
+            device - torch.device, either cuda or cpu
+            exclude_single - are we omitting sinlge cases, default is True
+            shuffle_case_order - shuffle cases, default is True
+            split_by_tumor_id - split the cases by tumorId, default is True
+
+
+    """
     def __init__(self,
                  doc_embeds,
                  Y,
+                 idxs,
                  tasks,
                  metadata,
[…]
