+++ fresco/run_clc.py
@@ -277,7 +277,7 @@ def run_case_level(args):
                                                dw.tasks,
                                                device)
 
-    dw.make_grouped_cases(outputs, clc_params.model_args, device, seed)
+    dw.make_grouped_cases(outputs, valid_params.model_args, device, seed)
 
     # 3. create a CLC model
     print("\nDefining a CLC model")
