### 2023-08-15 70a9022
Update run_clc.py
