### 2023-08-16 4ea39e1
update weights info in config.yml files
