### 2023-08-16 63bce5f
update clc example from code update
