+++ configs/clc_step2.yml
@@ -8,7 +8,7 @@ data_kwargs:
   shuffle_case_order: False
   subset_proportion: 1.0
   reproducible: True
-  random_seed: 42
+  random_seed: 56789
 
 # general training args
 train_kwargs:
