+++ paper/paper.md
@@ -9,6 +9,7 @@ tags:
 authors:
   - name: Adam Spannaus
     affiliation: 1
+    orcid: 0000-0002-2521-3657
     corresponding: True
   - name: John Gounley
     affiliation: 1
