### 2023-08-18 #5345 @aspannaus
Hi @jmschrei,

Thanks for all the assistance with this submission. I've completed the [additional author tasks](https://github.com/openjournals/joss-reviews/issues/5345#issuecomment-1679243196) and the manuscript is ready to go. Let us know if there's anything else you need from us. Thanks!
