### 2023-08-21 #5345 @aspannaus
@jmschrei, when you say 'explicitly include the license' do you mean the text of the license itself? The license is listed on the doi link. Thanks
