### 2023-09-04 #5345 @Kevin-Mattheus-Moerman
@aspannaus I am the AEiC for this track and here to help with final steps. Apologies for the delay, I was out of the office for a bit.

I have just looked at archive link provided: https://doi.org/10.11578/dc.20230227.2. That link does not seem to archive the actual software. Can you clarify if you've archived the software (full content of your software repository)? This is what is required for JOSS.
