+++ paper/shared/refs.bib
@@ -85,6 +85,7 @@
   title={Pyhealth: A python library for health predictive models},
   author={Zhao, Yue and Qiao, Zhi and Xiao, Cao and Glass, Lucas and Sun, Jimeng},
   journal={arXiv preprint arXiv:2101.04209},
+  doi = {10.48550/arXiv.2101.04209},
   year={2021}
 }
 
@@ -92,6 +93,7 @@
   title={MONAI: An open-source framework for deep learning in healthcare},
   author={Cardoso, M Jorge and Li, Wenqi and Brown, Richard and Ma, Nic and Kerfoot, Eric and Wang, Yiheng and Murrey, Benjamin and Myronenko, Andriy and Zhao, Can and Yang, Dong and others},
   journal={arXiv preprint arXiv:2211.02701},
+  doi={10.48550/arXiv.2211.02701},
   year={2022}
 }
 
@@ -122,6 +124,7 @@
   title={Ehrkit: A python natural language processing toolkit for electronic health record texts},
   author={Li, Irene and You, Keen and Tang, Xiangru and Qiao, Yujie and Huang, Lucas and Hsieh, Chia-Chun and Rosand, Benjamin and Radev, Dragomir},
   journal={arXiv preprint arXiv:2204.06604},
+  doi={10.48550/arXiv.2204.06604},
   year={2022}
 }
 
