### 2023-09-07 #5345 @Kevin-Mattheus-Moerman
@aspannaus I see the DOIs seem fixed. However, are you also able to address the issue with the archiving link? :point_up:
