+++ README.md
@@ -3,7 +3,9 @@
 ## Motivation
 The National Cancer Institute (NCI) monitors population level cancer trends as part of its Surveillance, Epidemiology, and End Results (SEER) program. This program consists of state or regional level cancer registries which collect, analyze, and annotate cancer pathology reports. From these annotated pathology reports, each individual registry aggregates cancer phenotype information and summary statistics about cancer prevalence to facilitate population level monitoring of cancer incidence. Extracting cancer phenotype from these reports is a labor intensive task, requiring specialized knowledge about the reports and cancer. Automating this information extraction process from cancer pathology reports has the potential to  improve not only the quality of the data by extracting information in a consistent manner across registries, but to improve the quality of patient outcomes by reducing the time to assimilate new data and enabling time-sensitive applications such as precision medicine. Here we present FrESCO: Framework for Exploring Scalable Computational Oncology, a modular deep-learning natural language processing (NLP) library for extracting pathology information from clinical text documents.
 
-Documentation is available over at [Read the Docs](https://doe-nci-fresco.readthedocs.io/en/latest/).
+Documentation is available over at [Read the Docs](https://doe-nci-fresco.readthedocs.io/en/latest/), and a software archive at Zenodo [![DOI](https://zenodo.org/badge/639999591.svg)](https://zenodo.org/badge/latestdoi/639999591).
+
+
 
 
 ## Quickstart Guide
