### 2023-09-11 #5345 @Kevin-Mattheus-Moerman
@aspannaus nearly there. The archive title should match the paper title `FrESCO: Framework for Exploring Scalable Computational Oncology`.
