### 2023-09-11 3adaecb
Create CITATION.cff
